Refactor code organization and separate concerns
