+++ lib/models/jmdict.ex
@@ -1,28 +1,42 @@
 defmodule JMDictEx.Models.JMDict do
+  @derive [Poison.Encoder]
+  @type t :: %__MODULE__{
+          common_only: boolean(),
+          dict_date: String.t(),
+          dict_revisions: [String.t()],
+          languages: [String.t()],
+          tags: %{String.t() => String.t()},
+          version: String.t(),
+          words: [Word.t()]
+        }
+  defstruct [:common_only, :dict_date, :dict_revisions, :languages, :tags, :version, :words]
+
   defmodule Gloss do
+    @derive [Poison.Encoder]
     @type t :: %__MODULE__{
-      gender: String.t() | nil,
-      lang: String.t(),
-      text: String.t(),
-      type: String.t() | nil
-    }
+            gender: String.t() | nil,
+            lang: String.t(),
+            text: String.t(),
+            type: String.t() | nil
+          }
     defstruct [:gender, :lang, :text, :type]
   end
 
   defmodule Sense do
+    @derive [Poison.Encoder]
     @type t :: %__MODULE__{
-      antonym: [String.t()],
-      applies_to_kana: [String.t()],
-      applies_to_kanji: [String.t()],
-      dialect: [String.t()],
-      field: [String.t()],
-      gloss: [Gloss.t()],
-      info: [String.t()],
-      language_source: [String.t()],
-      misc: [String.t()],
-      part_of_speech: [String.t()],
-      related: [[String.t()]]
-    }
+            antonym: [String.t()],
+            applies_to_kana: [String.t()],
+            applies_to_kanji: [String.t()],
+            dialect: [String.t()],
+            field: [String.t()],
+            gloss: [Gloss.t()],
+            info: [String.t()],
+            language_source: [String.t()],
+            misc: [String.t()],
+            part_of_speech: [String.t()],
+            related: [[String.t()]]
+          }
     defstruct [
       :antonym,
       :applies_to_kana,
@@ -39,59 +53,33 @@ defmodule JMDictEx.Models.JMDict do
   end
 
   defmodule Kana do
+    @derive [Poison.Encoder]
     @type t :: %__MODULE__{
-      applies_to_kanji: [String.t()],
-      common: boolean(),
-      tags: [String.t()],
-      text: String.t()
-    }
+            applies_to_kanji: [String.t()],
+            common: boolean(),
+            tags: [String.t()],
+            text: String.t()
+          }
     defstruct [:applies_to_kanji, :common, :tags, :text]
   end
 
   defmodule Kanji do
+    @derive [Poison.Encoder]
     @type t :: %__MODULE__{
-      common: boolean(),
-      text: String.t()
-    }
+            common: boolean(),
+            text: String.t()
+          }
     defstruct [:common, :text]
   end
 
   defmodule Word do
+    @derive [Poison.Encoder]
     @type t :: %__MODULE__{
-      id: String.t(),
-      kana: [Kana.t()],
-      kanji: [Kanji.t()],
-      sense: [Sense.t()]
-    }
+            id: String.t(),
+            kana: [Kana.t()],
+            kanji: [Kanji.t()],
+            sense: [Sense.t()]
+          }
     defstruct [:id, :kana, :kanji, :sense]
   end
-
-  @type t :: %__MODULE__{
-    common_only: boolean(),
-    dict_date: String.t(),
-    dict_revisions: [String.t()],
-    languages: [String.t()],
-    tags: %{String.t() => String.t()},
-    version: String.t(),
-    words: [Word.t()]
-  }
-  defstruct [:common_only, :dict_date, :dict_revisions, :languages, :tags, :version, :words]
-
-  def decode(json) do
-    Poison.decode(json,
-      as: %__MODULE__{
-        dict_revisions: [],
-        languages: [],
-        tags: %{},
-        words: [%Word{}]
-      }
-    )
-  end
-
-  # def decode!(json), do: Poison.decode!(json, as: %__MODULE__{
-  #   dict_revisions: [],
-  #   languages: [],
-  #   tags: %{},
-  #   words: [%Word{}]
-  # })
 end
